+++ src/MediatRItemExtension.Shared/Helpers/Operation/BindHandlerHelper.cs
@@ -177,9 +177,14 @@ namespace MediatRItemExtension.Helpers.Operation
                     model.HandlerHandleReturnValueName,
                     -1);
 
-                handler.AddParameter("request", model.OperationName, -1);
+                var requestPosition = model.OperationProcessing == ProcessType.Async ? 0 : -1;
+                handler.AddParameter("request", model.OperationName, requestPosition);
+
                 if (model.OperationProcessing == ProcessType.Async)
-                    handler.AddParameter("cancellationToken", "CancellationToken", -1);
+                {
+                    var cancellationTokenParam = handler.AddParameter("cancellationToken", "CancellationToken", -1);
+                    cancellationTokenParam.EndPoint.CreateEditPoint().Insert(" = default");
+                }
 
                 handler.StartPoint.CreateEditPoint().ReplaceText(0, model.HandlerAccessor,
                     (int)vsEPReplaceTextOptions.vsEPReplaceTextAutoformat);