+++ src/MediatRItemExtension.Shared/Helpers/Operation/BuildOperationHelper.cs
@@ -53,7 +53,6 @@ namespace MediatRItemExtension.Helpers.Operation
         {
             try
             {
-
                 ThreadHelper.ThrowIfNotOnUIThread();
                 var requestProjectItem = projectItems.AddFromTemplate(classTemplate, $"{model.OperationName}.cs");